refactor: create stats component
